requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
requests
pandas
openpyxl
aiofiles